test-requirements.txt
coverage>=4.0.3
nose>=1.3.7
pluggy>=0.3.1
py>=1.4.31
randomize>=0.13
pytest==7.4.3
allure-pytest==2.13.2
pytest-ordering==0.6
pytest-xdist==3.5.0
PyYAML==6.0.1
pytest-assume==2.4.3
pytest-asyncio==0.21.1
asyncio==3.4.3
pytest-tornasync>=0.6.0
pytest-trio==0.8.0
pytest-twisted==1.14.0
Twisted==23.10.0